File used in feature_importances.py
This is the first CSV file that Hongze uploaded. I have used it with my feature_importances.py file to assign the significance of each feature to every.

Note this file only contains data for a single pollinator, and is used to return the importances of different features relevant to that exact pollinator.